+++ docs/quadrature.pdf
@@ -7609,8 +7609,8 @@ endobj
 613 0 obj
 <<
   /Creator (Typst 0.14.2)
-  /ModDate (D:20260516045230Z)
-  /CreationDate (D:20260516045230Z)
+  /ModDate (D:20260516045409Z)
+  /CreationDate (D:20260516045409Z)
 >>
 endobj
 
@@ -7621,7 +7621,7 @@ endobj
   /Subtype /XML
 >>
 stream
-<?xpacket begin="﻿" id="W5M0MpCehiHzreSzNTczkc9d"?><x:xmpmeta xmlns:x="adobe:ns:meta/" x:xmptk="xmp-writer"><rdf:RDF xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"><rdf:Description rdf:about="" xmlns:dc="http://purl.org/dc/elements/1.1/"  xmlns:xmp="http://ns.adobe.com/xap/1.0/"  xmlns:xmpMM="http://ns.adobe.com/xap/1.0/mm/"  xmlns:xmpTPg="http://ns.adobe.com/xap/1.0/t/pg/"  xmlns:pdf="http://ns.adobe.com/pdf/1.3/" ><xmp:CreatorTool>Typst 0.14.2</xmp:CreatorTool><dc:language><rdf:Bag><rdf:li>en</rdf:li></rdf:Bag></dc:language><xmp:ModifyDate>2026-05-16T04:52:30+00:00</xmp:ModifyDate><xmp:CreateDate>2026-05-16T04:52:30+00:00</xmp:CreateDate><xmpTPg:NPages>7</xmpTPg:NPages><dc:format>application/pdf</dc:format><xmpMM:InstanceID>fsL2Gy05lFlGxYoNdPQzoQ==</xmpMM:InstanceID><xmpMM:DocumentID>fsL2Gy05lFlGxYoNdPQzoQ==</xmpMM:DocumentID><xmpMM:RenditionClass>proof</xmpMM:RenditionClass><pdf:PDFVersion>1.7</pdf:PDFVersion></rdf:Description></rdf:RDF></x:xmpmeta><?xpacket end="r"?>
+<?xpacket begin="﻿" id="W5M0MpCehiHzreSzNTczkc9d"?><x:xmpmeta xmlns:x="adobe:ns:meta/" x:xmptk="xmp-writer"><rdf:RDF xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"><rdf:Description rdf:about="" xmlns:dc="http://purl.org/dc/elements/1.1/"  xmlns:xmp="http://ns.adobe.com/xap/1.0/"  xmlns:xmpMM="http://ns.adobe.com/xap/1.0/mm/"  xmlns:xmpTPg="http://ns.adobe.com/xap/1.0/t/pg/"  xmlns:pdf="http://ns.adobe.com/pdf/1.3/" ><xmp:CreatorTool>Typst 0.14.2</xmp:CreatorTool><dc:language><rdf:Bag><rdf:li>en</rdf:li></rdf:Bag></dc:language><xmp:ModifyDate>2026-05-16T04:54:09+00:00</xmp:ModifyDate><xmp:CreateDate>2026-05-16T04:54:09+00:00</xmp:CreateDate><xmpTPg:NPages>7</xmpTPg:NPages><dc:format>application/pdf</dc:format><xmpMM:InstanceID>Mx3ICEwn5MJTt9dMxiEFbQ==</xmpMM:InstanceID><xmpMM:DocumentID>Mx3ICEwn5MJTt9dMxiEFbQ==</xmpMM:DocumentID><xmpMM:RenditionClass>proof</xmpMM:RenditionClass><pdf:PDFVersion>1.7</pdf:PDFVersion></rdf:Description></rdf:RDF></x:xmpmeta><?xpacket end="r"?>
 endstream
 endobj
 
@@ -8266,7 +8266,7 @@ trailer
   /Size 616
   /Root 615 0 R
   /Info 613 0 R
-  /ID [(fsL2Gy05lFlGxYoNdPQzoQ==) (fsL2Gy05lFlGxYoNdPQzoQ==)]
+  /ID [(Mx3ICEwn5MJTt9dMxiEFbQ==) (Mx3ICEwn5MJTt9dMxiEFbQ==)]
 >>
 startxref
 227328
